Improve library structure
