+++ package.json
@@ -9,9 +9,16 @@
     "metarhia",
     "lowcode",
     "nocode",
-    "javascript"
+    "javascript",
+    "markdown",
+    "dsl",
+    "bpmn"
   ],
   "main": "lowscript.js",
+  "files": [
+    "lib/",
+    "lowscript.d.ts"
+  ],
   "engines": {
     "node": "^12.9 || 14 || 16 || 17 || 18"
   },
@@ -34,12 +41,14 @@
     "fmt": "prettier --write \"**/*.js\" \"**/*.json\" \"**/*.md\" \"**/.*rc\""
   },
   "devDependencies": {
+    "@types/node": "^17.0.21",
     "eslint": "^8.14.0",
     "eslint-config-metarhia": "^7.0.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.0.0",
     "metatests": "^0.8.2",
-    "prettier": "^2.6.2"
+    "prettier": "^2.6.2",
+    "typescript": "^4.6.2"
   }
 }
