+++ test/examples/Store.md
